ACL update: eigenlijk is entitystring de Id van de acl
git-svn-id: https://svn.science.uu.nl/repos/project.fisme.java/DWOProject/DWO-LMS/DWOServer/trunk@23908 58dd2db4-0f11-0410-9df4-da15f00f50bd

+++ src/main/java/fi/dwo/server/PersistentDataManagers/core/ACLManager.java
@@ -8,6 +8,7 @@ import fi.dwo.server.persistence.DwoEmfFactory;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.persistence.EntityManager;
@@ -204,7 +205,14 @@ public class ACLManager {
         for (PersistentACL item: list) {
           boolean present = acls.stream().anyMatch(a -> item.getAclID().equals(a.getAclID()) );
           if (!present) {
-            em.remove(item);
+            Optional<PersistentACL> find = acls.stream().filter(a -> item.getEntity().equals(a.getEntity())).findAny();
+            if (find.isPresent())
+            {
+              find.get().setAclID(item.getAclID());
+              find.get().setOptlock(item.getOptlock());
+            }
+            else
+                em.remove(item);
           }
         }
         list = new ArrayList<>();
